Render html items on web
